Search: Move FOLDER_TITLE_HAPPY_PATH to folders.ts, delete search.ts
The constant was the sole remaining content in search.ts after the legacy
handler removal. It describes a folder title defined in folders.ts
DEFAULT_FOLDERS, so move it there and derive it from the source of truth.
Delete the now-empty search.ts file.
