--- public/app/features/alerting/unified/mocks/server/handlers/search.ts
@@ -1,1 +1,0 @@
-export const FOLDER_TITLE_HAPPY_PATH = 'Folder A';
